SigmaPi/requirements.txt
Django==1.8.7
djangorestframework>=3
argparse==1.2.1
beautifulsoup4==4.3.2
requests==2.0.1
wsgiref==0.1.2
django-sendfile==0.3.4
